cleanup: move to util
